Wrap survey choices in !unsafe option
We are already wrapping the default choice in an !unsafe option so we should
also do the choices. If not, when you try to import the resource the
Automation Controller API will complain about the default choice not being
in the list of choices.

The behavior will operate different if choice is an integer vs a string.